requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
flake8==3.9.2
flake8-html==0.4.1
importlib-metadata==4.5.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
zipp==3.4.1
